// libs/features/payment/src/lib/shared/models/repayment-details.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */
import { BilingualText, GosiCalendar } from '@gosi-ui/core';

export class RepaymentDetails {
  additionalPaymentDetails: string;
  amountTransferred: number;
  bankName: BilingualText;
  bankType: BilingualText;
  paymentMethod: BilingualText;
  paymentReferenceNo: number;
  receiptMode: BilingualText;
  transactionDate: GosiCalendar;
}